+++ assets/src/scripts/monitoring-location/components/hydrograph/index.test.js
@@ -326,11 +326,6 @@ describe('monitoring-location/components/hydrograph module', () => {
             expect(selectAll('#select-time-series').size()).toBe(1);
         });
 
-        it('should have tooltips for the select series table', () => {
-            // one for each of the two parameters and the WaterAlert links
-            expect(selectAll('table .usa-tooltip').size()).toBe(10);
-        });
-
         it('should have data tables for hydrograph data', () => {
             expect(select('#iv-hydrograph-data-table-container').size()).toBe(1);
             expect(select('#gw-hydrograph-data-table-container').size()).toBe(1);