+++ assets/src/scripts/monitoring-location/components/hydrograph/parameters.test.js
@@ -1,5 +1,6 @@
 import {select} from 'd3-selection';
 import sinon from 'sinon';
+import config from 'ui/config';
 
 import * as utils from 'ui/utils';
 
@@ -8,7 +9,7 @@ import * as hydrographData from 'ml/store/hydrograph-data';
 
 import * as dataIndicator from './data-indicator';
 import {TEST_HYDROGRAPH_PARAMETERS} from './mock-hydrograph-state';
-import {drawSelectionTable} from './parameters';
+import {drawSelectionList} from './parameters';
 
 describe('monitoring-location/components/hydrograph/parameters module', () => {
     utils.mediaQuery = jest.fn().mockReturnValue(true);
@@ -19,6 +20,32 @@ describe('monitoring-location/components/hydrograph/parameters module', () => {
            selectedDateRange: 'P7D',
            selectedParameterCode: '72019'
        }
+    };
+
+    config.ivPeriodOfRecord = {
+        '00060': {
+            begin_date: '1980-01-01',
+            end_date: '2020-01-01'
+        },
+        '72019': {
+            begin_date: '1980-04-01',
+            end_date: '2020-04-01'
+        },
+        '00010': {
+            begin_date: '1981-04-01',
+            end_date: '2019-04-01'
+        }
+
+    };
+    config.gwPeriodOfRecord = {
+        '72019': {
+            begin_date: '1980-03-31',
+            end_date: '2020-03-31'
+        },
+        '62610': {
+            begin_date: '1980-05-01',
+            end_date: '2020-05-01'
+        }
     };
 
     let div;
@@ -43,27 +70,26 @@ describe('monitoring-location/components/hydrograph/parameters module', () => {
         store = configureStore({
             hydrographParameters: {}
         });
-        drawSelectionTable(div, store, '11112222');
+        drawSelectionList(div, store, '11112222');
         expect(div.select('#select-time-series').size()).toBe(0);
     });
 
-    it('Expects the selection table to be rendered with the appropriate rows and selection', () => {
+    it('Expects the selection list to be rendered with the appropriate rows and selection', () => {
         store = configureStore(TEST_STATE);
-        drawSelectionTable(div, store, '11112222');
+        drawSelectionList(div, store, '11112222');
 
         const container = div.select('#select-time-series');
         expect(container.size()).toBe(1);
-        expect(container.select('table').size()).toBe(1);
-        expect(container.selectAll('tbody tr').size()).toBe(5);
-
-        expect(container.select('tbody input:checked').attr('value')).toEqual('72019');
+        expect(container.selectAll('.grid-row-container-row').size()).toBe(5);
+        expect(container.selectAll('.expansion-toggle').size()).toBe(8); // Note - there are two for each parameter with expansion rows
+        expect(container.select('input:checked').attr('value')).toEqual('72019');
     });
 
     it('Expects changing the selection retrieves hydrograph data', () => {
         store = configureStore(TEST_STATE);
-        drawSelectionTable(div, store, '11112222');
+        drawSelectionList(div, store, '11112222');
 
-        const rowOne = div.select('tbody tr:first-child');
+        const rowOne = div.select('#container-row-00060');
         rowOne.dispatch('click');
 
         expect(store.getState().hydrographState.selectedParameterCode).toEqual('00060');
@@ -77,5 +103,49 @@ describe('monitoring-location/components/hydrograph/parameters module', () => {
             loadCompare: false,
             loadMedian: false
         });
+    });
+
+    it('Expects clicking on a row will expand and contract the correct rows', function() {
+        store = configureStore(TEST_STATE);
+        drawSelectionList(div, store, '11112222');
+
+        const firstTargetRow = div.select('#container-row-00010');
+        const secondTargetRow = div.select('#container-row-72019');
+        expect(select('#expansion-container-row-00010').attr('hidden')).toBe('true');
+        expect(select('#expansion-container-row-72019').attr('hidden')).toBe('true');
+
+        firstTargetRow.dispatch('click');
+        expect(select('#expansion-container-row-00010').attr('hidden')).toBe(null);
+        expect(select('#expansion-container-row-72019').attr('hidden')).toBe('true');
+
+        secondTargetRow.dispatch('click');
+        expect(select('#expansion-container-row-00010').attr('hidden')).toBe('true');
+        expect(select('#expansion-container-row-72019').attr('hidden')).toBe(null);
+    });
+
+    it('Expects clicking the row toggle will expand the correct row and set the toggle', function() {
+        store = configureStore(TEST_STATE);
+        drawSelectionList(div, store, '11112222');
+
+        const firstToggleTarget = div.select('#expansion-toggle-desktop-00010');
+        const secondToggleTarget = div.select('#expansion-toggle-desktop-72019');
+        expect(firstToggleTarget.attr('aria-expanded')).toBe('false');
+        expect(select('#expansion-container-row-00010').attr('hidden')).toBe('true');
+        expect(select('#expansion-container-row-72019').attr('hidden')).toBe('true');
+
+        firstToggleTarget.dispatch('click');
+        expect(firstToggleTarget.attr('aria-expanded')).toBe('true');
+        expect(select('#expansion-container-row-00010').attr('hidden')).toBe(null);
+        expect(select('#expansion-container-row-72019').attr('hidden')).toBe('true');
+
+        secondToggleTarget.dispatch('click');
+        expect(secondToggleTarget.attr('aria-expanded')).toBe('true');
+        expect(select('#expansion-container-row-00010').attr('hidden')).toBe('true');
+        expect(select('#expansion-container-row-72019').attr('hidden')).toBe(null);
+
+        secondToggleTarget.dispatch('click'); // click same target a second time
+        expect(secondToggleTarget.attr('aria-expanded')).toBe('false');
+        expect(select('#expansion-container-row-00010').attr('hidden')).toBe('true');
+        expect(select('#expansion-container-row-72019').attr('hidden')).toBe('true');
     });
 });